requirements.txt.txt
aiogram==3.7.1
python-dotenv
